+++ odbc-api/src/buffers/columnar.rs
@@ -21,9 +21,9 @@ use odbc_sys::{CDataType, Date, Time, Timestamp};
 
 /// Since buffer shapes are same for all time / timestamps independent of the precision and we do
 /// not know the precise SQL type. In order to still be able to bind time / timestamp buffer as
-/// input without requiring the user to seperatly specify the precision, we declare Nano second
-/// precision, to not loose any information.
-const DEFAULT_TIME_PRECISION: i16 = 9;
+/// input without requiring the user to seperatly specify the precision, we declare 100 Nano second
+/// precision. This was the highest precision still supported by MSSQL in the tests.
+const DEFAULT_TIME_PRECISION: i16 = 7;
 
 /// A borrowed view on the valid rows in a column of a [`crate::buffers::ColumnarRowSet`].
 ///